st2common/in-requirements.txt
# Remeber to list implicit packages here, otherwise version won't be fixated!
apscheduler
python-dateutil
eventlet
jinja2
jsonschema
kombu
mongoengine
networkx
oslo.config
paramiko
pyyaml
pymongo
python-keyczar
requests
retrying
semver
six
tooz
bencode
ipaddr
git+https://github.com/StackStorm/pecan.git@st2-patched#egg=pecan
swagger_spec_validator
